group 1 refactor
