+++ cypress/support/index.js
@@ -16,23 +16,24 @@
 // Import commands.js using ES2015 syntax:
 import './commands';
 import '@cypress/code-coverage/support';
+import './api_functions';
 
 // Import Reusable Tests definitions
-import './reusable_tests/navigation_bar';
-import './reusable_tests/login';
-import './reusable_tests/create_account';
-import './reusable_tests/account_settings';
-import './reusable_tests/favorites';
-import './reusable_tests/footer_bar';
-import './reusable_tests/suggest_resource';
-import './reusable_tests/search_page';
-import './reusable_tests/announcement';
-import './reusable_tests/language';
-import './reusable_tests/utility';
-import './reusable_tests/details';
-import './reusable_tests/mexico_crime_map';
-import './reusable_tests/static_resource';
-import './reusable_tests/forgot_password';
+import './tests_definitions/navigation_bar';
+import './tests_definitions/login';
+import './tests_definitions/account_creation';
+import './tests_definitions/account_settings';
+import './tests_definitions/favorites';
+import './tests_definitions/footer_bar';
+import './tests_definitions/suggest_resource';
+import './tests_definitions/search_page';
+import './tests_definitions/announcement';
+import './tests_definitions/language';
+import './tests_definitions/utility';
+import './tests_definitions/details';
+import './tests_definitions/mexico_crime_map';
+import './tests_definitions/static_resource';
+import './tests_definitions/forgot_password';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
