+++ cypress/support/tests_definitions/details.js
