+++ cypress/support/tests_definitions/favorites.js
