+++ cypress/support/tests_definitions/footer_bar.js
