+++ cypress/support/tests_definitions/navigation_bar.js
