+++ cypress/support/tests_definitions/search_page.js
