+++ cypress/support/tests_definitions/static_resource.js
